requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
pydantic
python-multipart